Fix #81: GUI programs shouldn't exit on closing the parent terminal
Issue reproduction steps:

1. Spawn a GUI program e.g. open a PDF file in zathura or evince.
2. Without quitting `nnn` close the terminal.
3. Notice that the application quits too.

Can be a nagging issue if someone is not using a drop-down terminal.

Fix: detach a GUI child and start it in a new session.

There are 2 aspects to this commit:

- It fixes #81: in cases where we do not wait for a spawned child we can assume
that the child is a GUI process. We detach and spawn the child in a new session.
- It changes the behaviour to @Rahi374's suggestion in PR #40 i.e. invoke the DE
opener in F_NOWAIT mode.