Shorten home path in the header to "~".
In the UNIX world "~" is a synonym for the home folder of the current
user. This commit substitutes the full path of the home for the current
user with the char "~" in the header.

An user that prefers to see the whole address for the home directory
can do so passing the command line argument "F" or using the "F"
toggle when the software is already running.