Fix no file picked with NNN_TMPFILE set and quit with 'q'
If NNN_TMPFILE is exported, the selection buffer gets cleared as it is
combined with `^G` (clear picked files and exit). So even when 'q' is
pressed to quit the program the selection is lost. This was introduced
in commit d1d491c102f5da1bf6a2c8c21fa71530e74f90d3 after release v2.7.