Do not default to `xterm`
It makes more sense to default to the terminals people are less likely
to have. If a user has `xterm` installed, then that is, most probably,
because it ships with X11. Whereas if a user has `st` or `alacritty`
instlled, then that is because he/she chose to install it.