fix crash under _FORTIFY_SOURCE
when built with _FORTIFY_SOURCE it will check whether the buffer
has as much space as the argument passed to snprintf:

	https://github.com/bminor/glibc/blob/7b544224f82d20019f9b28522ebf8114a372d1a2/debug/snprintf_chk.c#L28-L29

this results in some false positives when the snprintf provided
len argument is bigger than the buffer size (but the result
would have fit into the buffer anyways).

fix this by passing the proper size to snprintf as argument.
(the +1 len isn't necessary, but add it just in case.)

Fixes: https://github.com/jarun/nnn/issues/1931